requirements.txt
click==7.1.1
cmake==3.16.3
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.18.2
opencv-python-headless==4.2.0.32
pi==0.1.2
Pillow==7.0.0
goprocam==4.0.3
